refactor(policies): integrates policy details into view
